refactor: separete hendle command
